SQL-1964: Move existing outdated test to separate file and refactor helper function
